check if fpgad.pid process is valid in scripts

+++ scripts/coverage-gtapi-mock-drv.sh
@@ -19,7 +19,9 @@ lcov -c -i -d . -o coverage.base
 LD_PRELOAD="$PWD/lib/libmock.so" ./bin/fpgad -d
 sleep 1
 CTEST_OUTPUT_ON_FAILURE=1 make test
-kill $(cat /tmp/fpgad.pid)
+if [ "$(pgrep fpgad)" -eq "$(cat /tmp/fpgad.pid)" ]; then
+	kill $(cat /tmp/fpgad.pid)
+fi
 deactivate
 
 find */**/opae-c.dir -iname "*.gcda" -exec chmod 664 '{}' \;
