+++ scripts/test-gtapi-mock-drv.sh
@@ -10,7 +10,9 @@ cmake .. -DBUILD_TESTS=ON -DCMAKE_BUILD_TYPE=Debug
 make mock gtapi fpgad pyopae
 LD_PRELOAD="$PWD/lib/libmock.so" ./bin/fpgad -d
 CTEST_OUTPUT_ON_FAILURE=1 make test
-kill $(cat /tmp/fpgad.pid)
+if [ "$(pgrep fpgad)" -eq "$(cat /tmp/fpgad.pid)" ]; then
+	kill $(cat /tmp/fpgad.pid)
+fi
 deactivate
 
 echo "test-gtapi-mock-drv build PASSED"
